Add formula selection functionality and make Controller to Singleton
